backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
spacy==3.5.0
Pillow
requests